refactor: extract generic Coqui bridge into coqui_config.py
voice_config_from_coqui had grown to handle GlowTTS, VITS (VitsCharacters) and
FastPitch -- it is a generic Coqui-TTS config bridge, not GlowTTS-specific. Move
it to phoonnx/engines/coqui_config.py; glowtts_config keeps only the Larynx
bridge and re-exports voice_config_from_coqui for back-compat. Also None-safe the
Graphemes/IPAPhonemes path (phonemes/characters may be null in some configs).
