+++ tests/test_fastpitch.py
@@ -62,7 +62,7 @@ def test_config_bridge_fastpitch_engine():
 
 def test_coqui_bridge_targets_fastpitch_engine():
     # the shared coqui bridge (also used by GlowTTS) can build a FastPitch config
-    from phoonnx.engines.glowtts_config import voice_config_from_coqui
+    from phoonnx.engines.coqui_config import voice_config_from_coqui
     cfg = {"use_phonemes": True, "phoneme_language": "en-us", "add_blank": False,
            "characters": {"characters": "abc", "punctuations": "!?.", "pad": "_",
                           "eos": "&", "bos": "*", "blank": "<BLNK>"}}
@@ -75,7 +75,7 @@ def test_coqui_vocab_is_sorted_like_coqui():
     # coqui's Graphemes/IPAPhonemes default is_sorted=True: the symbol set is
     # sorted alphabetically before ids are assigned. Regression guard for the
     # bug where the bridge kept config order -> right voice, wrong words.
-    from phoonnx.engines.glowtts_config import voice_config_from_coqui
+    from phoonnx.engines.coqui_config import voice_config_from_coqui
     cfg = {"use_phonemes": True, "phoneme_language": "en-us", "phonemizer": "gruut",
            "add_blank": False, "characters": {"phonemes": "zyxbac", "punctuations": "!?",
                                                "pad": "_", "eos": "~", "bos": "^"}}
@@ -88,7 +88,7 @@ def test_coqui_vocab_is_sorted_like_coqui():
 
 
 def test_coqui_phonemizer_honors_config_field():
-    from phoonnx.engines.glowtts_config import voice_config_from_coqui
+    from phoonnx.engines.coqui_config import voice_config_from_coqui
     from phoonnx.config import PhonemeType
     base = {"use_phonemes": True, "characters": {"phonemes": "abc", "pad": "_", "eos": "~", "bos": "^"}}
     assert voice_config_from_coqui({**base, "phonemizer": "gruut"}, lang_code="en").phoneme_type == PhonemeType.GRUUT
@@ -98,7 +98,7 @@ def test_coqui_phonemizer_honors_config_field():
 def test_coqui_vits_characters_vocab():
     # VITS uses VitsCharacters: [pad] + punct + graphemes + ipa + [blank], NOT
     # sorted, is_unique=False (blank id = full-list length, not deduped length).
-    from phoonnx.engines.glowtts_config import voice_config_from_coqui
+    from phoonnx.engines.coqui_config import voice_config_from_coqui
     from phoonnx.config import Engine
     cfg = {"use_phonemes": True, "phonemizer": "espeak",
            "characters": {"characters_class": "TTS.tts.models.vits.VitsCharacters",
